refactor: Filter 컴포넌트 로직 분리 및 usePromotionNotification 훅 도입
