+++ frontend/src/pages/PromotionPage/PromotionListPage.tsx
@@ -3,6 +3,7 @@ import Header from '@/components/common/Header/Header';
 import { PAGE_VIEW } from '@/constants/eventName';
 import useTrackPageView from '@/hooks/Mixpanel/useTrackPageView';
 import { useGetPromotionArticles } from '@/hooks/Queries/usePromotion';
+import usePromotionNotification from '@/hooks/Queries/usePromotionNotification';
 import isInAppWebView from '@/utils/isInAppWebView';
 import Filter from '../../components/common/Filter/Filter';
 import PromottionGrid from './components/list/PromotionGrid/PromotionGrid';
@@ -12,12 +13,13 @@ const PromotionListPage = () => {
   useTrackPageView(PAGE_VIEW.PROMOTION_LIST_PAGE);
 
   const { data, isLoading, isError } = useGetPromotionArticles();
+  const hasNotification = usePromotionNotification();
 
   return (
     <>
       <Header hideOn={['webview']} />
       <Styled.Container>
-        {!isInAppWebView() && <Filter />}
+        {!isInAppWebView() && <Filter hasNotification={hasNotification} />}
         <Styled.Wrapper>
           {isLoading && <p>로딩 중...</p>}
           {isError && <p>오류가 발생했습니다.</p>}
